Add IExecutionData interface and move all usage to the interface.
Abstracts the underlying execution data implementation to allow for more
complex execution data storage mechanisms.
For offline instrumentation, classes store the IExecutionData object and
make calls into it to set probe data. This results in additional
overhead of looking up the vtable for the interface and making the
method call, rather than just directly setting the value in the boolean
array.

Bug: 147904124
Test: m EMMA_INSTRUMENT=true EMMA_INSTRUMENT_FRAMEWORK=true and verify
Java coverage collection and parsing

Change-Id: I41235626b1040a6a21c6aef300f2dfe064393e1f